fix: add executeMergeEvidence to structured output and move evidence tests
- executeMerge now populates executeMergeEvidence in structuredContent
  when evidence is on (was missing, text-only before)
- dryRunAll now merges duplicatePairs, mergeSuggestions, themeGroups,
  and relationshipClusters from delegated strategies into its structured
  output (was text-only before)
- Add ConsolidateExecuteMergeEvidence type and Zod schema
- Move evidence integration tests from sync-migrations to
  memory-lifecycle where they belong
- Assert both text and structured output in execute-merge evidence test
- Assert evidence:false produces undefined executeMergeEvidence
