+++ tests/memory-lifecycle.integration.test.ts
@@ -16,6 +16,8 @@ import {
 
 import { writeFile } from "fs/promises";
 
+import { ConsolidateResultSchema } from "../src/structured-content.js";
+
 describe("memory-lifecycle", () => {
   it("supports global remember and forget with git disabled", async () => {
     const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
@@ -920,4 +922,151 @@ describe("memory-lifecycle", () => {
       await embeddingServer.close();
     }
   }, 35000);
+
+  it("returns consolidate evidence payloads for detect-duplicates and suggest-merges", async () => {
+    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
+    tempDirs.push(vaultDir);
+    const embeddingServer = await startFakeEmbeddingServer();
+
+    try {
+      const noteAText = await callLocalMcp(vaultDir, "remember", {
+        title: "Evidence merge note A",
+        content: "Merge evidence payload test for duplicate and merge suggestion analysis.",
+        tags: ["evidence", "merge"],
+        lifecycle: "permanent",
+        summary: "Create note A for consolidate evidence",
+        scope: "global",
+      }, embeddingServer.url);
+      const noteAId = extractRememberedId(noteAText);
+
+      const noteBText = await callLocalMcp(vaultDir, "remember", {
+        title: "Evidence merge note B",
+        content: "Merge evidence payload test for duplicate and merge suggestion analysis.",
+        tags: ["evidence", "merge"],
+        lifecycle: "temporary",
+        role: "research",
+        summary: "Create note B for consolidate evidence",
+        scope: "global",
+      }, embeddingServer.url);
+      const noteBId = extractRememberedId(noteBText);
+
+      await callLocalMcp(vaultDir, "relate", {
+        fromId: noteAId,
+        toId: noteBId,
+        type: "supersedes",
+        bidirectional: false,
+      }, embeddingServer.url);
+
+      const dupResponse = await callLocalMcpResponse(vaultDir, "consolidate", {
+        strategy: "detect-duplicates",
+        threshold: 0,
+        evidence: true,
+      }, embeddingServer.url);
+      const dupParsed = ConsolidateResultSchema.parse(dupResponse.structuredContent);
+      expect(dupParsed.duplicatePairs?.length).toBeGreaterThan(0);
+      expect(dupParsed.duplicatePairs?.[0]?.noteA.mergeRisk).toMatch(/low|medium|high/);
+      expect(dupParsed.duplicatePairs?.[0]?.mergeRisk).toMatch(/low|medium|high/);
+      expect(dupResponse.text).toContain("Merge risk:");
+
+      const suggestResponse = await callLocalMcpResponse(vaultDir, "consolidate", {
+        strategy: "suggest-merges",
+        threshold: 0,
+        evidence: true,
+      }, embeddingServer.url);
+      const suggestParsed = ConsolidateResultSchema.parse(suggestResponse.structuredContent);
+      expect(suggestParsed.mergeSuggestions?.length).toBeGreaterThan(0);
+      expect(suggestParsed.mergeSuggestions?.[0]?.notes[0]?.ageDays).toBeTypeOf("number");
+      expect(suggestParsed.mergeSuggestions?.[0]?.mergeRisk).toMatch(/low|medium|high/);
+      expect(suggestResponse.text).toContain("Evidence:");
+    } finally {
+      await embeddingServer.close();
+    }
+  }, 15000);
+
+  it("renders execute-merge evidence by default and suppresses it when evidence is false", async () => {
+    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
+    tempDirs.push(vaultDir);
+    const embeddingServer = await startFakeEmbeddingServer();
+
+    try {
+      const noteAText = await callLocalMcp(vaultDir, "remember", {
+        title: "Execute merge evidence A",
+        content: "First note for execute-merge evidence test.",
+        tags: ["evidence", "execute-merge"],
+        lifecycle: "permanent",
+        summary: "Create note A for execute-merge evidence",
+        scope: "global",
+      }, embeddingServer.url);
+      const noteAId = extractRememberedId(noteAText);
+
+      const noteBText = await callLocalMcp(vaultDir, "remember", {
+        title: "Execute merge evidence B",
+        content: "Second note for execute-merge evidence test.",
+        tags: ["evidence", "execute-merge"],
+        lifecycle: "temporary",
+        role: "research",
+        summary: "Create note B for execute-merge evidence",
+        scope: "global",
+      }, embeddingServer.url);
+      const noteBId = extractRememberedId(noteBText);
+
+      const mergeWithEvidence = await callLocalMcpResponse(vaultDir, "consolidate", {
+        strategy: "execute-merge",
+        evidence: true,
+        mergePlan: {
+          sourceIds: [noteAId, noteBId],
+          targetTitle: "Execute merge evidence target",
+        },
+      }, embeddingServer.url);
+
+      expect(mergeWithEvidence.text).toContain("Evidence:");
+      expect(mergeWithEvidence.text).toContain("risk:");
+      expect(mergeWithEvidence.text).toContain("Merge risk:");
+      expect(mergeWithEvidence.text).toMatch(/temporary, research/);
+      expect(mergeWithEvidence.text).toMatch(/permanent, untyped/);
+
+      const mergeParsed = ConsolidateResultSchema.parse(mergeWithEvidence.structuredContent);
+      expect(mergeParsed.executeMergeEvidence).toBeDefined();
+      expect(mergeParsed.executeMergeEvidence?.notes.length).toBe(2);
+      expect(mergeParsed.executeMergeEvidence?.notes[0]?.ageDays).toBeTypeOf("number");
+      expect(mergeParsed.executeMergeEvidence?.mergeRisk).toMatch(/low|medium|high/);
+
+      const noteCText = await callLocalMcp(vaultDir, "remember", {
+        title: "Execute merge no-evidence C",
+        content: "Third note for execute-merge without evidence.",
+        tags: ["evidence", "execute-merge-no-evidence"],
+        lifecycle: "permanent",
+        summary: "Create note C for execute-merge no-evidence",
+        scope: "global",
+      }, embeddingServer.url);
+      const noteCId = extractRememberedId(noteCText);
+
+      const noteDText = await callLocalMcp(vaultDir, "remember", {
+        title: "Execute merge no-evidence D",
+        content: "Fourth note for execute-merge without evidence.",
+        tags: ["evidence", "execute-merge-no-evidence"],
+        lifecycle: "permanent",
+        summary: "Create note D for execute-merge no-evidence",
+        scope: "global",
+      }, embeddingServer.url);
+      const noteDId = extractRememberedId(noteDText);
+
+      const mergeWithoutEvidence = await callLocalMcpResponse(vaultDir, "consolidate", {
+        strategy: "execute-merge",
+        evidence: false,
+        mergePlan: {
+          sourceIds: [noteCId, noteDId],
+          targetTitle: "Execute merge no-evidence target",
+        },
+      }, embeddingServer.url);
+
+      expect(mergeWithoutEvidence.text).not.toContain("Evidence:");
+      expect(mergeWithoutEvidence.text).not.toContain("Merge risk:");
+
+      const noEvidenceParsed = ConsolidateResultSchema.parse(mergeWithoutEvidence.structuredContent);
+      expect(noEvidenceParsed.executeMergeEvidence).toBeUndefined();
+    } finally {
+      await embeddingServer.close();
+    }
+  }, 15000);
 });
