+++ tests/sync-migrations.integration.test.ts
@@ -15,7 +15,6 @@ import {
 } from "./helpers/mcp.js";
 
 import {
-  ConsolidateResultSchema,
   GetResultSchema,
   ListResultSchema,
   MemoryGraphResultSchema,
@@ -289,144 +288,6 @@ describe("sync-migrations", () => {
       const after = await readFile(embeddingPath, "utf-8");
       expect(after).not.toBe("");
       expect(before).not.toBe("");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("returns consolidate evidence payloads for detect-duplicates and suggest-merges", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const noteAText = await callLocalMcp(vaultDir, "remember", {
-        title: "Evidence merge note A",
-        content: "Merge evidence payload test for duplicate and merge suggestion analysis.",
-        tags: ["evidence", "merge"],
-        lifecycle: "permanent",
-        summary: "Create note A for consolidate evidence",
-        scope: "global",
-      }, embeddingServer.url);
-      const noteAId = extractRememberedId(noteAText);
-
-      const noteBText = await callLocalMcp(vaultDir, "remember", {
-        title: "Evidence merge note B",
-        content: "Merge evidence payload test for duplicate and merge suggestion analysis.",
-        tags: ["evidence", "merge"],
-        lifecycle: "temporary",
-        role: "research",
-        summary: "Create note B for consolidate evidence",
-        scope: "global",
-      }, embeddingServer.url);
-      const noteBId = extractRememberedId(noteBText);
-
-      await callLocalMcp(vaultDir, "relate", {
-        fromId: noteAId,
-        toId: noteBId,
-        type: "supersedes",
-        bidirectional: false,
-      }, embeddingServer.url);
-
-      const dupResponse = await callLocalMcpResponse(vaultDir, "consolidate", {
-        strategy: "detect-duplicates",
-        threshold: 0,
-        evidence: true,
-      }, embeddingServer.url);
-      const dupParsed = ConsolidateResultSchema.parse(dupResponse.structuredContent);
-      expect(dupParsed.duplicatePairs?.length).toBeGreaterThan(0);
-      expect(dupParsed.duplicatePairs?.[0]?.noteA.mergeRisk).toMatch(/low|medium|high/);
-      expect(dupParsed.duplicatePairs?.[0]?.mergeRisk).toMatch(/low|medium|high/);
-      expect(dupResponse.text).toContain("Merge risk:");
-
-      const suggestResponse = await callLocalMcpResponse(vaultDir, "consolidate", {
-        strategy: "suggest-merges",
-        threshold: 0,
-        evidence: true,
-      }, embeddingServer.url);
-      const suggestParsed = ConsolidateResultSchema.parse(suggestResponse.structuredContent);
-      expect(suggestParsed.mergeSuggestions?.length).toBeGreaterThan(0);
-      expect(suggestParsed.mergeSuggestions?.[0]?.notes[0]?.ageDays).toBeTypeOf("number");
-      expect(suggestParsed.mergeSuggestions?.[0]?.mergeRisk).toMatch(/low|medium|high/);
-      expect(suggestResponse.text).toContain("Evidence:");
-    } finally {
-      await embeddingServer.close();
-    }
-  }, 15000);
-
-  it("renders execute-merge evidence by default and suppresses it when evidence is false", async () => {
-    const vaultDir = await mkdtemp(path.join(os.tmpdir(), "mnemonic-mcp-vault-"));
-    tempDirs.push(vaultDir);
-    const embeddingServer = await startFakeEmbeddingServer();
-
-    try {
-      const noteAText = await callLocalMcp(vaultDir, "remember", {
-        title: "Execute merge evidence A",
-        content: "First note for execute-merge evidence test.",
-        tags: ["evidence", "execute-merge"],
-        lifecycle: "permanent",
-        summary: "Create note A for execute-merge evidence",
-        scope: "global",
-      }, embeddingServer.url);
-      const noteAId = extractRememberedId(noteAText);
-
-      const noteBText = await callLocalMcp(vaultDir, "remember", {
-        title: "Execute merge evidence B",
-        content: "Second note for execute-merge evidence test.",
-        tags: ["evidence", "execute-merge"],
-        lifecycle: "temporary",
-        role: "research",
-        summary: "Create note B for execute-merge evidence",
-        scope: "global",
-      }, embeddingServer.url);
-      const noteBId = extractRememberedId(noteBText);
-
-      const mergeWithEvidence = await callLocalMcp(vaultDir, "consolidate", {
-        strategy: "execute-merge",
-        evidence: true,
-        mergePlan: {
-          sourceIds: [noteAId, noteBId],
-          targetTitle: "Execute merge evidence target",
-        },
-      }, embeddingServer.url);
-
-      expect(mergeWithEvidence).toContain("Evidence:");
-      expect(mergeWithEvidence).toContain("risk:");
-      expect(mergeWithEvidence).toContain("Merge risk:");
-      expect(mergeWithEvidence).toMatch(/temporary, research/);
-      expect(mergeWithEvidence).toMatch(/permanent, untyped/);
-
-      const noteCText = await callLocalMcp(vaultDir, "remember", {
-        title: "Execute merge no-evidence C",
-        content: "Third note for execute-merge without evidence.",
-        tags: ["evidence", "execute-merge-no-evidence"],
-        lifecycle: "permanent",
-        summary: "Create note C for execute-merge no-evidence",
-        scope: "global",
-      }, embeddingServer.url);
-      const noteCId = extractRememberedId(noteCText);
-
-      const noteDText = await callLocalMcp(vaultDir, "remember", {
-        title: "Execute merge no-evidence D",
-        content: "Fourth note for execute-merge without evidence.",
-        tags: ["evidence", "execute-merge-no-evidence"],
-        lifecycle: "permanent",
-        summary: "Create note D for execute-merge no-evidence",
-        scope: "global",
-      }, embeddingServer.url);
-      const noteDId = extractRememberedId(noteDText);
-
-      const mergeWithoutEvidence = await callLocalMcp(vaultDir, "consolidate", {
-        strategy: "execute-merge",
-        evidence: false,
-        mergePlan: {
-          sourceIds: [noteCId, noteDId],
-          targetTitle: "Execute merge no-evidence target",
-        },
-      }, embeddingServer.url);
-
-      expect(mergeWithoutEvidence).not.toContain("Evidence:");
-      expect(mergeWithoutEvidence).not.toContain("Merge risk:");
     } finally {
       await embeddingServer.close();
     }
